fix(Xover callback): fix potential crashes due to accessing mmap()-ed object invalidated via ftruncate()
The object in the file is deleted and truncated out using rec_del(),
but the pointer to the corresponding object in the memory is also invalidated
because the pointer to the object is mmap()-ed to the object in the file,
and the changes of the file is synchronized with the mmap()-ed memory.

The solution is to back up the object before deleting it with rec_del().

* maple/gem.c:
    * gem_log(): Change the type of parameter `hdr` to `const HDR *`
    * gem_delete()
    * gem_move()
* maple/mail.c:
    * mbox_delete()
* maple/myfavorite.c:
    * myfavorite_delete()
    * myfavorite_mov()
* maple/post.c:
    * post_delete() (not changed):
        * Needs no change since rec_del() is invoked with lazy_delete(),
          which does not actually delete the object
    * post_clean_delete()
* so/chatmenu.c:
    * chat_move()
* so/cleanrecommend.c:
    * cleanrecommend_clean()
* so/mailgem.c:
    * mailgem_move()
* so/personal.c:
    * personal_deny()