feat(log): message logger with log level (thanks @PichuChen)
Usage: `LOGF_*(&g_logger, "format", format_args...)`

g_logger.path and g_logger.lv_skip is initialized in maple/bbsd.c,
while g_logger.file is NULL and ignored by loggerf().

This makes loggerf() open the log file for writing
every time it is invoked.

Messages with its log level >= g_logger.lv_skip will be ignored,
so that g_logger.lv_skip can be modified at runtime to control
which messages should be output to the log file.

An extra newline will be appended to the message for output,
therefore it is not needed to include a trailing newline
in the message format string.

* include/logger.h:
    * Add enum LogLevel
    * Add the struct Logger
* lib/logger.c:
    * Implement loggerf()

* include/dao.h:
    * Add declaration for loggerf()
    * Add macros LOGF_*() for loggerf() shorthands

* lib/CMakeLists.txt & lib/Makefile:
    * Add logger.c to build target

* include/global.h:
    * Add declaration for the global logger g_logger
* maple/bbsd.c:
    * Add definition for global logger g_logger