fix(lib/passwd.c): fix wrongly generated SHA-256 signature since 8c67043c9925e01af60aa4c1dbe66e2c77720a58
* gensignature(): fix truncated SHA-256 hash

In 8c67043c9925e01af60aa4c1dbe66e2c77720a58,
where the SHA-256 signature was implemented,
gensignature() was made to put a '\0' at the new string end
of the generated signature.

However, the index for the new string end was wrong.

Due to the confusion of the string length and the string index
at the time of coding,
an erroneous `-1` was added to the index for the new string end,
and thus the last byte of the string was truncated.

To avoid similar logic errors, it should be remembered that,
if str_c is the concatenation of str_a and str_b,
`str_c[strlen(str_a) + strlen(str_b)] == '\0'` should hold.

* gensignature() & chksignature: fix straying '$' left in the SHA-256 hash

PASSLEN and PASSHASHLEN are both buffer sizes,
and thus adding `- 1` to them is needed to get the string length.

However, in the aforementioned commit,
the string length was confused at the time of coding.

As a result,
the straying '$' along with other unnecessary bytes are copied in gensignature()
and is unintentionally ignored in chksignature().

This fixes the above issues.