fix(makefiles): fix missing architecture info for x86 platforms for generating version info
The makefile variable MULTIARCH is used for generating version info,
however it was unset on x86 platforms.

Before 538da2caac923b2c6279600df0bb399e13f148f0, where the version info generation was implemented,
MULTIARCH was only used for setting the search paths for pkg-config on
x86_64 platforms to solve the package problems,
and therefore it was unnecessary to set MULTIARCH for x86 platforms.

However, since the aforementioned commit,
MULTIARCH is required even on x86 platforms for generating the version info,
but unfortunately it was ignored
and made the architecture info missing from the generated version info.

This fix solve this issue by ensuring that MULTIARCH is set on x86 platforms.