fix(UI/Xover): fix corrupted header after xo_thread() since 4940e01306d24fb244154383e9bfaee20c3898f1
Before commit 4940e01306d24fb244154383e9bfaee20c3898f1,
xo_thread() returned the following values
for the result of searching related articles:
* XO_NONE, when none are found,
* -1, when any is found,
* XO_BODY, when any is found and a page flip is needed.

The code handling xo_thread() in xover() (now a part of xover_key())
cleared the cursor only when -1 was returned.

In that commit, xo_thread() was changed to return the following values
to handle footer redrawing:
* XO_NONE, when none are found,
* XO_FOOT, when none are found and the footer needs to be redrawn,
* `XO_MOVE + XO_REL`, when any is found,
* `XR_BODY + XO_MOVE + XO_REL`, when any is found and a page flip is needed.

However, in that commit, the aforementioned code was erroneously changed
to clear the cursor whenever a related article is found,
regardless of any page flips.

This made the cursor clearing occur on the lines outside of the Xover list,
and thus sometimes corrupted the displayed header and necker on the screen.

This commit fixes the issue
by reintroducing the check against page flips to the aforementioned code.