fix(dns_mx()): fix only 1 MX entry can be returned; return as many as possible
MAXDNAME is larger than MAX_MXLIST (1024) in recent glibc,
where MAXDNAME is defined as 1025.

This made the calculated pointer value for determing the end of the MX list,
which was `mxlist + MAX_MXLIST - MAXDNAME - 2`,
to be in front of the MX list string pointer,
and thus prevented more MX entries from being appended to the list.

The issue is fixed by refining the handling of the end of the MX list
so that the calculated pointer value for determing the end of the list
becomes `mxlist + mxlist_sz - 2`.

* Add parameter mxlist_sz to dns_mx() for the buffer size of parameter `mxlist`
* Use the remaining buffer size of `mxlist` instead of the fixed size MAXDNAME
  for invoking dn_expand()
* Add variable mxlist_end to drop the reusing of variable `domain`
* Check whether the string end of MX entries reach or exceed mxlist_end
  before appending the list separator
* Drop an awkward linewrap before a function call argument list