fix(MENU): fix loading of DlFuncArg obj broken since 3acbab2dac434dd2219731b691627d1635f42217
This issue caused crashes when invoking a DlFuncArg menu item
for the 3nd time while hotswapping was enabled
or for the next time after a failed dynamic loading
regardless of whether hotswapping was enabled.

This is because the member `DlFuncArg::func` of the menu item object
was wrongly unconditionally replaced by the result of `DL_GET()`.

- Cases where the 1st loading successed while hotswapping was enabled:

Event    | Accessed field                              | Resulting field
 ---     | ---                                         | ---
Init     | (None)                                      | `DlFuncArg::func` (DL obj name str)
1st load | `DlFuncArg::func` (valid DL obj name str)   | `FuncArg::func` (function pointer)
2nd load | `DlFuncArg::func` (invalid DL obj name str) | `FuncArg::func` (`NULL`)
3rd load | `DlFuncArg::func` (NULL-dereferencing)      | (undefined)

Note that the `MENU::umode` (determines the accessed field) for the menu item
was not updated due to the way hotswapping works.

- Cases where the 1st loading failed, regardless of whether hotswapping was enabled:

Event    | Accessed field                              | Resulting field
 ---     | ---                                         | ---
Init     | (None)                                      | `DlFuncArg::func` (DL obj name str)
1st load | `DlFuncArg::func` (valid; failed to load)   | `FuncArg::func` (`NULL`)
2nd load | `DlFuncArg::func` (NULL-dereferencing)      | (undefined)

This issue affected menu_reset[], the "系統重置" (resetting system) menu.

This issue is fixed by copying the `DlFuncArg` object
to avoid unwanted modifications to the pointed `DlFuncArg` object,
which should not be modified if hotswapping is enabled.

* maple/menu.c: domenu_exec() & maple/popupmenu.c: do_cmd():
    * Move local var `mitem` into the new helper union `MItem m`