fix(bbsd): prevent telnet_init() from dropping IAC cmds
When IAC cmds was sent by the client
before or during telnet_init() was invoked,
these cmds were all discarded, including the NAWS cmd.

This made clients which sends the NAWS cmd
as soon as the connection is established
unable to initialized the screen dimensions of the connection.

This issue is fixed by handling IAC cmds with iac_process() instead.

* telnet_init() & term_init():
    * Remove input data handling codes
* term_init():
    * Merge svr[] and the data sending code into telnet_init()
    * Only set up the screen dimension with the default values