fix(xo_thread()): fix RS_FIRST became RS_SECOND in 441c0eb662f2d464f98fa88613104fa5db6096d5
This issue is fixed by using `goto found;` when a matching article is found
instead of setting `match` on matching and examining `match` later.

The algorithm before that commit:
- If an exact match is found, break the loop and use `pos` as the result.
- Else, if RS_FIRST is set and `near` is valid (>= 0), use `near` as the result.
- Else not found.

The wrong algorithm introduced by that commit:
- If RS_FIRST is set and `near` is valid (>= 0), use `near` as the result.
- Else, if an exact match is found, break the loop and use `pos` as the result.
- Else not found.

Since `near` is the farthest result until the exact match,
this wrong algorithm produces the second farthest result (if any)
when an exact match is found.

The new algorithm introduced by this commit:
- If a matching is found during the loop, `goto found;`
- Else, if RS_FIRST is set and `near` is valid (>= 0), use `near` as the result
  and then `goto found;`.
- Else do not `goto found;`.