fix(pmore mf_parseHeaders()): fix crash when line 0 contains >= 2GiB chars
This would make the following malloc()
try to allocate a memory region of roughly 16EiB due to integer overflow,
which would certainly fail on any contemporary average computer.

As a result, malloc() would return NULL
and later crash the program due to NULL-dereferencing.

Though this case is prevented by the in-BBS editor.

* use ptrdiff_t instead of `int` for local `l`