docs(pfterm): clarify the swapped usage of scrolling up/down
"Scrolling up/down" usually refers to the movement of the screen range.

However, its usages are swapped in pfterm.

This terminology is easily confused,
since the content moves in the opposite direction to the screen range
and mobile users scroll by swiping in the same direction as the content.

To avoid further confusions,
the moving direction of the screen range is now hinted with "forward/backward",
while the swapped usages are kept as-is.